+++ Models/Terreno.cs
@@ -14,7 +14,7 @@
         public required int Azimuth { get; set; }
         public bool InstalacionEstructura { get; set; }
         public double? Separacion { get; set; }
-        public int? PotenciaTotal { get; set; }
+        public double? PotenciaTotal { get; set; }
         public int? TotalPaneles { get; set; }
     }
 }
